Tezrisat_Backend/requirements.txt
Django~=4.2.16
django-cors-headers==4.4.0
djangorestframework~=3.15.2
python-dotenv~=1.0.1
whitenoise==6.9.0

dj_database_url==3.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.10

langchain==0.3.23
langchain-chroma==0.1.4
langchain-community==0.3.0
langchain-core==0.3.51
langchain-openai==0.3.3
langchain-text-splitters==0.3.8
langgraph==0.2.27
langgraph-checkpoint==1.0.12
langsmith==0.1.147

trafilatura~=1.12.2
urllib3==2.2.1
requests~=2.32.3
requests-ntlm==1.2.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0

tqdm~=4.66.4
numpy==1.26.4
pypdf==4.2.0
pylatexenc==2.10
sympy==1.12
pyparsing==3.1.2
regex==2024.11.6
python-dateutil==2.9.0.post0
PyPika==0.48.9
rsa==4.9
cryptography==42.0.5

uvicorn==0.34.0
watchdog==4.0.1
watchfiles==1.0.4
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1

wikipedia==1.4.0
wolframalpha==5.1.3

google-auth==2.38.0
google-search-results==2.4.2
googleapis-common-protos==1.69.2
